+++ db/aggregate-test.js
@@ -1,11 +1,11 @@
 const { default: mongoose } = require('mongoose')
-const getShowtimesForCinemaGroupByDate = require('./aggregations/showtimes-by-date')
+const getShowtimesForCinemaGroupByMovie = require('./aggregations/showtimes-by-movie')
 
 let cinemaId = '623b47d48f8dbba189a1a6cb'
 
 async function execute() {
   await require('./index')
-  const showtimes = await getShowtimesForCinemaGroupByDate(cinemaId)
+  const showtimes = await getShowtimesForCinemaGroupByMovie(cinemaId)
 
   console.log(showtimes[0])
